+++ source/spi_api.c
@@ -36,7 +36,7 @@ static uint32_t spi_event_check(spi_t *obj);
 static void spi_buffer_tx_write(spi_t *obj);
 static void spi_buffer_rx_read(spi_t *obj);
 static void spi_enable_event_flags(spi_t *obj, uint32_t event, uint8_t enable);
-static void spi_buffer_set(spi_t *obj, void *tx, uint32_t tx_length, void *rx, uint32_t rx_length, uint8_t bit_width);
+static void spi_buffer_set(spi_t *obj, void *tx, size_t tx_length, void *rx, size_t rx_length, uint8_t bit_width);
 
 #define SPI_TX_FIFO_SIZE (4)
 #define SPI_RX_FIFO_SIZE (4)
@@ -184,7 +184,7 @@ static void spi_master_enable_interrupt(spi_t *obj, uint8_t enable)
     DSPI_HAL_SetRxFifoDrainDmaIntMode(obj->spi.address, kDspiGenerateIntReq, enable);
 }
 
-void spi_master_transfer(spi_t *obj, void *tx, uint32_t tx_length, void *rx, uint32_t rx_length, uint8_t bit_width, uint32_t handler, uint32_t event, DMAUsage hint)
+void spi_master_transfer(spi_t *obj, void *tx, size_t tx_length, void *rx, size_t rx_length, uint8_t bit_width, uint32_t handler, uint32_t event, DMAUsage hint)
 {
     if (hint != DMA_USAGE_NEVER && obj->spi.dma_state == DMA_USAGE_ALLOCATED) {
         // setup dma done, activate
@@ -411,7 +411,7 @@ uint8_t spi_active(spi_t *obj)
     }
 }
 
-static void spi_buffer_set(spi_t *obj, void *tx, uint32_t tx_length, void *rx, uint32_t rx_length, uint8_t bit_width)
+static void spi_buffer_set(spi_t *obj, void *tx, size_t tx_length, void *rx, size_t rx_length, uint8_t bit_width)
 {
     obj->tx_buff.buffer = tx;
     obj->tx_buff.length = tx_length;
